refactoring the entire listings with search and filter
